+++ OpenBench/workload.py
@@ -18,191 +18,39 @@
 #                                                                             #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-# Module serves a singular purpose, to invoke: get_workload(Machine)
-# Selects a Test, from the active tests, with the following criteria...
-#
-# 1. Remove all tests with Engines not supported by the Worker
-# 2. Remove all tests with unmet Syzygy requirements
-# 3. Remove all tests with thread-odds, if the Worker is hyperthreading
-# 4. Remove all tests where our machine will exceed the Worker Limit
-# 5. Remove all tests where our machine will exceed the Thread Limit
-# 6. Remove all tests that don't have the highest priority in the test list
-#
-# At this point, we select from these candidate tests using the following:
-#
-# 1. Randomly select from any tests that have 0 workers
-# 2. Repeat the same test ( if it exists ), if distribution is still "Fair"
-# 3. Select the test with the most "Unfair" distribution of workers
+import OpenBench.views
 
-import random
-import sys
+from OpenBench.models import *
 
-import OpenBench.utils
+def create_workload(request, workload_type):
 
-from OpenBench.config import OPENBENCH_CONFIG
-from OpenBench.models import Result, Test
+    assert workload_type in [ 'TEST', 'TUNE' ]
 
-def get_workload(machine):
+    if not request.user.is_authenticated:
+        return OpenBench.views.redirect(request, '/login/', error='Only enabled users can create tests')
 
-    # Select a workload from the possible ones, if we can
-    if not (test := select_workload(machine)):
-        return {}
+    if not Profile.objects.get(user=request.user).enabled:
+        return OpenBench.views.redirect(request, '/login/', error='Only enabled users can create tests')
 
-    # Fetch or create the Result object for the test
-    try: result = Result.objects.get(test=test, machine=machine)
-    except: result = Result(test=test, machine=machine)
+    if request.method == 'GET':
+        data     = { 'networks' : list(Network.objects.all().values()) }
+        template = 'create_test.html' if workload_type == 'TEST' else 'create_tune.html'
+        return OpenBench.views.render(request, template, data)
 
-    # Update the Machine's status and save everything
-    machine.workload = test.id; machine.save(); result.save()
-    return { 'workload' : workload_to_dictionary(test, result) }
+    test, errors = OpenBench.utils.create_new_test(request)
+    if errors != [] and errors != None:
+        return OpenBench.views.redirect(request, '/newTest/', error='\n'.join(errors))
 
-def select_workload(machine):
+    if warning := OpenBench.utils.branch_is_out_of_date(test):
+        warning = 'Consider Rebasing: Dev (%s) appears behind Base (%s)' % (test.dev.name, test.base.name)
 
-    # Find valid workloads, given the current distribution of workers
-    active, distribution = worker_distribution(machine)
-    if not (tests := filter_valid_workloads(active, machine, distribution)):
-        return {}
+    username = request.user.username
+    profile  = Profile.objects.get(user=request.user)
+    summary  = 'CREATE P=%d TP=%d' % (test.priority, test.throughput)
+    LogEvent.objects.create(author=username, summary=summary, log_file='', test_id=test.id)
 
-    # Find the tests most deserving of resources currently
-    ratios = [distribution[x.id]['threads'] / x.throughput for x in tests]
-    lowest_indices = [i for i, r in enumerate(ratios) if r == min(ratios)]
+    if not OpenBench.config.USE_CROSS_APPROVAL and profile.approver:
+        test.approved = True; test.save()
 
-    # Machine is out of date; or there is an unassigned test
-    if machine.workload not in distribution or min(ratios) == 0:
-        return tests[random.choice(lowest_indices)]
+    return OpenBench.views.redirect(request, '/index/', warning=warning)
 
-    # Determine the "fair" ratio given the total threads and total throughput of tests
-    total_threads    = sum([distribution[x.id]['threads'] for x in tests])
-    total_throughput = sum([x.throughput for x in tests])
-    fair_ratio       = total_threads / total_throughput
-
-    # Repeat the same test if the distribution is still fair
-    if min(ratios) / fair_ratio > 0.75:
-        return Test.objects.get(id=machine.workload)
-
-    # Fallback to simply doing the least attention given test
-    return tests[random.choice(lowest_indices)]
-
-def worker_distribution(machine):
-
-    # For each test ( which some worker claims to be running ), return
-    # a dict with the # of workers and # of threads actively on the test
-
-    tests = OpenBench.utils.get_active_tests()
-
-    distribution = {
-        test.id : { 'workers' : 0, 'threads' : 0 } for test in tests
-    }
-
-    # Don't count ourselves; and don't include non-active tests
-    for x in OpenBench.utils.getRecentMachines():
-        if machine != x and x.workload in distribution:
-            distribution[x.workload]['workers'] += 1
-            distribution[x.workload]['threads'] += x.info['concurrency']
-
-    return tests, distribution
-
-def filter_valid_workloads(tests, machine, distribution):
-
-    # Skip engines that the Machine cannot handle
-    for engine in OPENBENCH_CONFIG['engines'].keys():
-        if engine not in machine.info['supported']:
-            tests = tests.exclude(dev_engine=engine)
-            tests = tests.exclude(base_engine=engine)
-
-    # Skip tests with unmet Syzygy requirements
-    for K in range(machine.info['syzygy_max'] + 1, 10):
-        tests = tests.exclude(syzygy_adj='%d-MAN' % (K))
-        tests = tests.exclude(syzygy_wdl='%d-MAN' % (K))
-
-    # Skip tests that would waste available Threads or exceed them
-    threads      = machine.info['concurrency']
-    sockets      = machine.info['sockets']
-    hyperthreads = machine.info['physical_cores'] < threads
-    options = [x for x in tests if valid_assignment(machine, x, distribution)]
-
-    # Finally refine for tests of the highest priority
-    if not options: return []
-    highest_prio = max(options, key=lambda x: x.priority).priority
-    return [test for test in options if test.priority == highest_prio]
-
-def valid_assignment(machine, test, distribution):
-
-    # Extract thread requirements from the test itself
-    dev_threads  = int(OpenBench.utils.extract_option(test.dev_options,  'Threads'))
-    base_threads = int(OpenBench.utils.extract_option(test.base_options, 'Threads'))
-
-    # Extract the information from our machine
-    threads      = machine.info['concurrency']
-    sockets      = machine.info['sockets']
-    hyperthreads = machine.info['physical_cores'] < threads
-
-    # Refuse if there are not enough threads-per-socket for the test
-    if max(dev_threads, base_threads) > (threads / sockets):
-        return False
-
-    # Refuse thread odds if we are using hyperthreads
-    if dev_threads != base_threads and hyperthreads:
-        return False
-
-    # Refuse to assign more workers than the test will allow
-    current_workers = distribution[test.id]['workers'] if test.id in distribution else 0
-    if test.worker_limit and current_workers >= test.worker_limit:
-        return False
-
-    # Refuse to assign more threads than the test will allow
-    current_threads = distribution[test.id]['threads'] if test.id in distribution else 0
-    if test.thread_limit and current_threads + threads > test.thread_limit:
-        return False
-
-    # All Criteria have been met
-    return True
-
-def workload_to_dictionary(test, result):
-
-    return {
-
-        'result' : {
-            'id'  : result.id
-        },
-
-        'test' : {
-            'id'            : test.id,
-            'syzygy_wdl'    : test.syzygy_wdl,
-            'syzygy_adj'    : test.syzygy_adj,
-            'win_adj'       : test.win_adj,
-            'draw_adj'      : test.draw_adj,
-            'workload_size' : test.workload_size,
-            'book'          : OPENBENCH_CONFIG['books'][test.book_name],
-
-            'dev' : {
-                'id'           : test.dev.id,
-                'name'         : test.dev.name,
-                'source'       : test.dev.source,
-                'sha'          : test.dev.sha,
-                'bench'        : test.dev.bench,
-                'engine'       : test.dev_engine,
-                'options'      : test.dev_options,
-                'network'      : test.dev_network,
-                'time_control' : test.dev_time_control,
-                'nps'          : OPENBENCH_CONFIG['engines'][test.dev_engine]['nps'],
-                'build'        : OPENBENCH_CONFIG['engines'][test.dev_engine]['build'],
-                'private'      : OPENBENCH_CONFIG['engines'][test.dev_engine]['private'],
-            },
-
-           'base' : {
-                'id'           : test.base.id,
-                'name'         : test.base.name,
-                'source'       : test.base.source,
-                'sha'          : test.base.sha,
-                'bench'        : test.base.bench,
-                'engine'       : test.base_engine,
-                'options'      : test.base_options,
-                'network'      : test.base_network,
-                'time_control' : test.base_time_control,
-                'nps'          : OPENBENCH_CONFIG['engines'][test.base_engine]['nps'],
-                'build'        : OPENBENCH_CONFIG['engines'][test.base_engine]['build'],
-                'private'      : OPENBENCH_CONFIG['engines'][test.base_engine]['private'],
-            },
-        },
-    }
