+++ Templates/OpenBench/base.html
@@ -86,6 +86,7 @@
             <ul>
                 <li class="header">Actions</li>
                 <li><a href="/newTest"><i class="fa-solid fa-plus fa-fw"></i> Create Test</a></li>
+                <li><a href="/newTune"><i class="fa-solid fa-plus fa-fw"></i> Create Tune</a></li>
                 <li><a href="/newNetwork"><i class="fa-solid fa-upload fa-fw"></i> Upload Net</a></li>
             </ul>
