+++ Templates/OpenBench/create_tune.html
@@ -1,0 +1,159 @@
+{% extends "OpenBench/base.html" %}
+
+{% load static %}
+
+{% block scripts %}
+
+    {{ config|json_script:"json-config" }}
+    {{ networks|json_script:"json-networks" }}
+    {{ profile.repos|json_script:"json-repos" }}
+
+    <script src="{% static 'create_tune.js' %}"></script>
+
+    <script>
+        document.addEventListener(
+            'DOMContentLoaded', function () {
+                set_engine('{{ profile.engine }}');
+                enforce_default_text('repo', 'https://github.com/');
+            }, false
+        );
+    </script>
+{% endblock %}
+
+{% block content %}
+
+    <form method="POST" action="/newTest/">
+        {% csrf_token %}
+
+        <div class="form w-100">
+            <div class="col">
+
+                <h3> Engine </h3>
+
+                <div class="row">
+                    <label> Engine </label>
+                    <select id="engine" name="engine" onchange="set_engine(value);">
+                        {% for engine in config.engines %}
+                            <option value="{{ engine }}">{{ engine }}</option>
+                        {% endfor %}
+                    </select>
+                </div>
+                <div class="row">
+                    <label> Source </label> <input id="repo" name="repo">
+                </div>
+
+                <h3> Engine Settings </h3>
+
+                <div class="row">
+                    <label> Branch </label> <input id="branch" name="branch" autofocus>
+                </div>
+                <div class="row">
+                    <label> Bench </label> <input value="Autofill" name="bench">
+                </div>
+                <div class="row">
+                    <label> Network </label> <select id="network" name="network"></select>
+                </div>
+                <div class="row">
+                    <label> Options </label> <input id="options" name="options">
+                </div>
+                <div class="row">
+                    <label> Time </label> <input id="time_control" name="time_control">
+                </div>
+
+                <h3> SPSA Input </h3>
+
+                <div class="row">
+
+                    <textarea style="margin-left:24px" id="spsa_input" name="spsa_input" rows=17></textarea>
+                </div>
+
+            </div>
+
+            <div class="col">
+
+                <h3> SPSA Settings </h3>
+
+                <div class="row">
+                    <label> Opening Book </label> <select id="book_name" name="book_name">
+                    {% for name, data in config.books.items %}
+                        <option value="{{ name }}">{{ name }}</option>
+                    {% endfor %}
+                </select>
+                </div>
+
+                <div class="row">
+                    <label> Alpha </label> <input value="0.602" id="spsa_alpha" name="spsa_alpha">
+                </div>
+                <div class="row">
+                    <label> Gamma </label> <input value="0.101" id="spsa_gamma" name="spsa_gamma">
+                </div>
+                <div class="row">
+                    <label> A-Ratio </label> <input value="0.1" id="spsa_A_ratio" name="spsa_A_ratio">
+                </div>
+                <div class="row">
+                    <label> Iterations </label> <input value="200000" id="spsa_iterations" name="spsa_iterations">
+                </div>
+                <div class="row">
+                    <label> Pairs-Per </label> <input value="1" id="spsa_pairs_per" name="spsa_pairs_per">
+                </div>
+
+                <h3> General Settings </h3>
+                <div class="row">
+                    <label> Priority </label> <input value="1" name="priority">
+                </div>
+                <div class="row">
+                    <label> Throughput </label> <input value="1000" id="throughput" , name="throughput">
+                </div>
+                <div class="row">
+                    <label> Worker Limit</label> <input value="5" id="worker_limit" name="worker_limit">
+                </div>
+                <div class="row">
+                    <label> Thread Limit </label> <input value="None" id="thread_limit" name="thread_limit">
+                </div>
+
+                <div class="row">
+                    <label> Syzygy WDL </label> <select name="syzygy_wdl">
+                        <option selected value="DISABLED"> Disabled </option>
+                        <option value="OPTIONAL"> Optional </option>
+                        <option value="7-MAN"> 7-Man </option>
+                        <option value="6-MAN"> 6-Man </option>
+                        <option value="5-MAN"> 5-Man </option>
+                        <option value="4-MAN"> 4-Man </option>
+                        <option value="3-MAN"> 3-Man </option>
+                    </select>
+                </div>
+
+                <h3> Adjudication Settings </h3>
+                <div class="row">
+                    <label> Syzygy ADJ. </label>
+                    <select name="syzygy_adj">
+                        <option value="DISABLED"> Disabled </option>
+                        <option selected value="OPTIONAL"> Optional </option>
+                        <option value="7-MAN"> 7-Man </option>
+                        <option value="6-MAN"> 6-Man </option>
+                        <option value="5-MAN"> 5-Man </option>
+                        <option value="4-MAN"> 4-Man </option>
+                        <option value="3-MAN"> 3-Man </option>
+                    </select>
+                </div>
+                <div class="row">
+                    <label> Win ADJ. </label> <input id="win_adj" name="win_adj">
+                </div>
+                <div class="row">
+                    <label> Draw ADJ. </label> <input id="draw_adj" name="draw_adj">
+                </div>
+            </div>
+        </div>
+
+        <input type="submit" class="anchorbutton btn-start mt-3 mw-84" name="submit" value="Create SPSA Tune">
+    </form>
+
+    <div class="form">
+        <div class="col-full">
+            <div id="test-mode-buttons">
+                <!-- Generated based on the selected Engine -->
+            </div>
+        </div>
+    </div>
+
+{% endblock %}
